Don't try to send throug closed socket in NodeDiscovery
When stopping the services, and PeerManager is still alive after
NodeDiscovery was stopped, it may try to send some discovery messages.
This will cause DatagramServer.sendto to throw AttributeError, because
a closed DatagramServer doesn't have a `socket` attribute.

Prevent this, by checking if the DatagramServer is closed, and dropping
the message if it is.